Fix exception when enabling widget-selection mode in Widget Inspector...
when popover is open.

Visual Studio Code's Widget Inspector causes PopoverItem to throw an
exception when trying to recalculate the bounds of the injected BuildContext,
because the context has been unmounted possibly due to the popover being open
and the currently active route. For whatever reason, the widget is mounted
when didChangeDependencies is called, but not in the post-frame callback
that tries to use the deactivated context.